Speed up api (for multiple predictions)
Compute enrichements before predictions
