+++ detect_ai_content/ml_logic/for_texts/using_ml_features/features/features.py
@@ -181,22 +181,3 @@ def BERT_encode(tokenizer, text_sentence, add_special_tokens=True):
     input_ids = torch.tensor([tokenizer.encode(text_sentence, add_special_tokens=add_special_tokens)]).to(device)
     mask_idx = torch.where(input_ids == tokenizer.mask_token_id)[1].tolist()[0]
     return input_ids, mask_idx
-
-from nltk.tokenize import word_tokenize, sent_tokenize
-import textstat
-
-def enrich_lexical_diversity_readability(data):
-    enriched_data = data.copy()
-
-    # Lexical diversity (Unique words / Total words)
-    enriched_data['lexical_diversity'] = data['text'].apply(lambda x: len(set(word_tokenize(x))) / len(word_tokenize(x)) if len(word_tokenize(x)) > 0 else 0)
-
-    # Readability Scores
-    enriched_data['flesch_reading_ease'] = data['text'].apply(textstat.flesch_reading_ease)
-    enriched_data['smog_index'] = data['text'].apply(textstat.smog_index)
-    enriched_data['flesch_kincaid_grade'] = data['text'].apply(textstat.flesch_kincaid_grade)
-
-    # Some metrics
-    enriched_data['avg_word_length'] = data['text'].apply(lambda x: np.mean([len(word) for word in word_tokenize(x)]))
-
-    return enriched_data
